sensors/presence/requirements.txt
kafka
paho-mqtt